backend/requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Cache & Redis
django-redis>=5.4,<6.0
redis>=5.0,<6.0

# Authentication
PyJWT>=2.8,<3.0
cryptography>=41.0,<43.0
python-jose>=3.3,<4.0

# Image handling
Pillow>=10.0,<11.0
django-imagekit>=5.0,<6.0

# API utilities
drf-spectacular>=0.27,<1.0
django-extensions>=3.2,<4.0

# Production server
gunicorn>=21.0,<23.0
whitenoise>=6.6,<7.0

# Environment
python-dotenv>=1.0,<2.0
python-decouple>=3.8,<4.0

# Testing
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0
faker>=22.0,<24.0

# Code quality
black>=23.12,<25.0
isort>=5.13,<6.0
flake8>=7.0,<8.0

# Utilities
python-slugify>=8.0,<9.0

# FSM (Finite State Machine)
django-fsm>=2.8,<3.0
